Initial commit for python-lspconfig.nvim
This is being split out from my nvim config as I suspect it might be useful for
others. The idea here is to auto configure `lspconfig` for `pyright`/`basedpyright`
based on a helper for finding the python path based on the root directory the
LSP found.

Users must opt in to having things auto configured, just set `opts = {}`.
Additionally there is a `python_path(dir)` helper in case you want to configure
things yourself.

Changing the active `VIRTUAL_ENV` is not in scope (at least for now). But
thanks to the helper you could always wire this up with the right config.